The Starry Night Update
GameModules folder now contains the Shader Factory Dynamic Sky. The sky is working 100% with Light probes and Reflection Probes.

SampleScene now has light and reflection probes and it is night time now. The skybox was set to night as a proof of concept for what things should change and how to have the scene look good and functional.

Next step will be adding the ability for the Dynamic Sky to turn from day to night in the material. And also have it's valued controlled by a system.